pyproject.toml
[project]
name = "webtoon-comprehension-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.9.0",
    "langchain==0.1.11",
    "langchain-openai==0.0.8",
    "langgraph==0.0.26",
    "pillow==10.2.0",
    "python-dotenv==1.0.1",
    "streamlit==1.32.2",
]
